Fix the bug of compile onnx model to openvino: both input_shape and batch were provided
Closes #430
